+++ customvectordrawable/src/main/AndroidManifest.xml
@@ -1,5 +1,5 @@
 <?xml version="1.0" encoding="utf-8"?>
 <manifest xmlns:android="http://schemas.android.com/apk/res/android"
-    package="com.suenara.customvectordrawable">
+    package="com.suenara.enhancedvectordrawable">
 
 </manifest>
