--- customvectordrawable/src/main/kotlin/com/suenara/customvectordrawable/AnimationTarget.kt
@@ -1,6 +1,0 @@
-package com.suenara.customvectordrawable
-
-import androidx.annotation.Keep
-
-@Keep
-interface AnimationTarget
