--- customvectordrawable/src/main/kotlin/com/suenara/customvectordrawable/VectorPathContainer.kt
@@ -1,5 +1,0 @@
-package com.suenara.customvectordrawable
-
-interface VectorPathContainer {
-    fun findPath(name: String): VectorPath?
-}
