+++ customvectordrawable/src/main/kotlin/com/suenara/enhancedvectordrawable/VectorAnimationContainer.kt
@@ -1,0 +1,8 @@
+package com.suenara.enhancedvectordrawable
+
+import android.animation.Animator
+
+interface VectorAnimationContainer : VectorPathContainer {
+    fun findAnimations(targetName: String): Animator?
+    fun invalidateAnimations()
+}
