+++ customvectordrawable/src/main/kotlin/com/suenara/enhancedvectordrawable/VectorPathContainer.kt
@@ -1,0 +1,5 @@
+package com.suenara.enhancedvectordrawable
+
+interface VectorPathContainer {
+    fun findPath(name: String): VectorPath?
+}
